Fix fatal float precision bug
Fix calculations with floats in ISR causing crash due to the lack of FPU
in the interrupt on the ESP32. Remove unnecessary double precision in
configuration variables and update tests. Fix clang-tidy again.

+++ src/utils/configuration.h
@@ -37,10 +37,10 @@ public:
     static const gpio_num_t ledPin = static_cast<gpio_num_t>(LED_PIN);
     static const bool isRgb = Configurations::ledPin == GPIO_NUM_NC ? false : IS_RGB;
     static const unsigned char impulsesPerRevolution = IMPULSES_PER_REVOLUTION;
-    static constexpr double flywheelInertia = FLYWHEEL_INERTIA;
+    static constexpr float flywheelInertia = FLYWHEEL_INERTIA;
     static const unsigned short ledBlinkFrequency = LED_BLINK_FREQUENCY;
-    static constexpr double sprocketRadius = SPROCKET_RADIUS / 100;
-    static constexpr double concept2MagicNumber = CONCEPT_2_MAGIC_NUMBER;
+    static constexpr float sprocketRadius = SPROCKET_RADIUS / 100;
+    static constexpr float concept2MagicNumber = CONCEPT_2_MAGIC_NUMBER;
 
     static constexpr Configurations::precision angularDisplacementPerImpulse = (2 * PI) / Configurations::impulsesPerRevolution;
     static const unsigned char driveHandleForcesMaxCapacity = UCHAR_MAX;
@@ -50,18 +50,18 @@ public:
     static const unsigned int rowingStoppedThresholdPeriod = ROWING_STOPPED_THRESHOLD_PERIOD * 1'000;
 
     // Drag factor filter settings
-    static constexpr double goodnessOfFitThreshold = GOODNESS_OF_FIT_THRESHOLD;
+    static constexpr float goodnessOfFitThreshold = GOODNESS_OF_FIT_THRESHOLD;
     static const unsigned int maxDragFactorRecoveryPeriod = MAX_DRAG_FACTOR_RECOVERY_PERIOD * 1'000;
-    static constexpr double lowerDragFactorThreshold = LOWER_DRAG_FACTOR_THRESHOLD / 1e6;
-    static constexpr double upperDragFactorThreshold = UPPER_DRAG_FACTOR_THRESHOLD / 1e6;
+    static constexpr float lowerDragFactorThreshold = LOWER_DRAG_FACTOR_THRESHOLD / 1e6;
+    static constexpr float upperDragFactorThreshold = UPPER_DRAG_FACTOR_THRESHOLD / 1e6;
     static const unsigned char dragCoefficientsArrayLength = DRAG_COEFFICIENTS_ARRAY_LENGTH;
 
     // Stroke phase detection filter settings
     static constexpr StrokeDetectionType strokeDetectionType = STROKE_DETECTION;
-    static constexpr double minimumPoweredTorque = MINIMUM_POWERED_TORQUE;
-    static constexpr double minimumDragTorque = MINIMUM_DRAG_TORQUE;
-    static constexpr double minimumRecoverySlopeMargin = MINIMUM_RECOVERY_SLOPE_MARGIN / 1e6;
-    static constexpr double minimumRecoverySlope = MINIMUM_RECOVERY_SLOPE;
+    static constexpr float minimumPoweredTorque = MINIMUM_POWERED_TORQUE;
+    static constexpr float minimumDragTorque = MINIMUM_DRAG_TORQUE;
+    static constexpr float minimumRecoverySlopeMargin = MINIMUM_RECOVERY_SLOPE_MARGIN / 1e6;
+    static constexpr float minimumRecoverySlope = MINIMUM_RECOVERY_SLOPE;
     static const unsigned int minimumRecoveryTime = MINIMUM_RECOVERY_TIME * 1'000;
     static const unsigned int minimumDriveTime = MINIMUM_DRIVE_TIME * 1'000;
     static const unsigned char impulseDataArrayLength = IMPULSE_DATA_ARRAY_LENGTH;
@@ -77,8 +77,8 @@ public:
     // Device power management settings
     static const gpio_num_t batteryPinNumber = BATTERY_PIN_NUMBER;
     static const unsigned char voltageDividerRatio = VOLTAGE_DIVIDER_RATIO;
-    static constexpr double batteryVoltageMin = BATTERY_VOLTAGE_MIN / Configurations::voltageDividerRatio;
-    static constexpr double batteryVoltageMax = BATTERY_VOLTAGE_MAX / Configurations::voltageDividerRatio;
+    static constexpr float batteryVoltageMin = BATTERY_VOLTAGE_MIN / Configurations::voltageDividerRatio;
+    static constexpr float batteryVoltageMax = BATTERY_VOLTAGE_MAX / Configurations::voltageDividerRatio;
     static const unsigned char batteryLevelArrayLength = BATTERY_LEVEL_ARRAY_LENGTH;
     static const unsigned char initialBatteryLevelMeasurementCount = INITIAL_BATTERY_LEVEL_MEASUREMENT_COUNT;
     static const unsigned int batteryMeasurementFrequency = BATTERY_MEASUREMENT_FREQUENCY * 60 * 1'000;
